rename heartbeat service file
