+++ packages/core/src/v3/utils/interval.ts
